refactor(v3.2.3): 抽 479 行机构级建模渲染到 lib/report/institutional
- trap_color_emoji + 10 个 _render_xxx (DCF/Comps/LBO/IC/catalyst/competitive/style/data_gap)
- 行数: 1708 → 1229 行 (-28% · 累计从 2964 -59%)
- 332 tests 全过 · e2e 零影响
